daisy: overlay: Disable vendor mismatch warning
* Remove the vendor mismatch message due to custom kernel

For Non Treble devices when build fingerprint is modified, At the initial boot Android throws up the error vendor image does not match with the system.
Setting to false by default, as no one would like to see that message.
Sometimes this annoying message appears depending on the kernel you want to use or change.

Change-Id: 8a1e2f5ec3bf87c7755b329a470ac24631727827
Signed-off-by: TogoFire <italomellopereira@gmail.com>
Signed-off-by: lupesoltec <dev@lupesoltec.com.br>